factorize msm_kzg part
